Re-instated smug reactions.
